refactor: 필수동의 컴포넌트 회원가입 폴더 안으로 이동
